backend/requirements_simple.txt
fastapi==0.115.6
uvicorn==0.32.1
pydantic==2.10.4
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.57.0
python-dotenv==1.0.0 